Extract mentor syncing to Git::SyncsMentors
