+++ test/services/git/syncs_website_copy_test.rb
@@ -1,7 +1,7 @@
 require_relative '../../test_helper'
 
 class Git::SyncsWebsiteCopyTest < ActiveJob::TestCase
-  test "creates mentors" do
+  test "syncs mentors" do
     Git::WebsiteContent.stubs(:repo_url).returns("file://#{Rails.root}/test/fixtures/website-copy")
     track = create(:track, slug: "go")
 
@@ -14,36 +14,5 @@ class Git::SyncsWebsiteCopyTest < ActiveJob::TestCase
     assert_equal "Katrina Owen", mentor.name
     assert_nil mentor.avatar_url
     assert_equal "kytrinyx", mentor.github_username
-    assert_nil mentor.link_text
-    assert_nil mentor.link_url
-    assert_equal "Bio", mentor.bio
-  end
-
-  test "updates mentors" do
-    Git::WebsiteContent.stubs(:repo_url).returns("file://#{Rails.root}/test/fixtures/website-copy")
-    track = create(:track, slug: "go")
-    mentor = create(:mentor,
-                    track: track,
-                    github_username: "kytrinyx",
-                    avatar_url: "avatars.com/avatar.png")
-
-    stub_repo_cache! do
-      Git::SyncsWebsiteCopy.()
-    end
-
-    mentor.reload
-    assert_nil mentor.avatar_url
-  end
-
-  test "deletes mentors" do
-    Git::WebsiteContent.stubs(:repo_url).returns("file://#{Rails.root}/test/fixtures/website-copy")
-    track = create(:track, slug: "go")
-    create(:mentor, github_username: "other_mentor")
-
-    stub_repo_cache! do
-      Git::SyncsWebsiteCopy.()
-    end
-
-    assert_equal 1, Mentor.count
   end
 end
